+++ packages/sync-rules/src/sync_plan/serialize.ts
@@ -1,6 +1,7 @@
 import { ParameterLookupScope } from '../HydrationState.js';
-import { TablePattern } from '../TablePattern.js';
+import { ImplicitSchemaTablePattern, TablePattern } from '../TablePattern.js';
 import { SqlExpression } from './expression.js';
+import { MapSourceVisitor, visitExpr } from './expression_visitor.js';
 import {
   ColumnSource,
   ColumnSqlParameterValue,
@@ -16,7 +17,10 @@ import {
   StreamOptions,
   StreamParameterIndexLookupCreator,
   StreamQuerier,
-  SyncPlan
+  SyncPlan,
+  TableProcessor,
+  TableProcessorTableValuedFunction,
+  TableProcessorTableValuedFunctionOutput
 } from './plan.js';
 
 /**
@@ -31,13 +35,38 @@ export function serializeSyncPlan(plan: SyncPlan): SerializedSyncPlanUnstable {
   const bucketIndex = new Map<StreamBucketDataSource, number>();
   const parameterIndex = new Map<StreamParameterIndexLookupCreator, number>();
   const expandingLookups = new Map<ExpandingLookup, LookupReference>();
+  const addedTableValuedFunctions = new Map<TableProcessorTableValuedFunction, number>();
 
-  function serializeTablePattern(pattern: TablePattern): SerializedTablePattern {
+  const replaceFunctionReferenceWithIndex = new MapSourceVisitor<
+    ColumnSqlParameterValue | TableProcessorTableValuedFunctionOutput,
+    ColumnSqlParameterValue | SerializedTableProcessorTableValuedFunctionOutput
+  >((value) => {
+    if ('function' in value) {
+      return { function: addedTableValuedFunctions.get(value.function)!, outputName: value.outputName };
+    } else {
+      return value;
+    }
+  });
+
+  function serializeTablePattern(pattern: ImplicitSchemaTablePattern): SerializedTablePattern {
     return {
       connection: pattern.connectionTag,
       schema: pattern.schema,
       table: pattern.tablePattern
     };
+  }
+
+  function serializeTableValued(source: TableProcessor): TableProcessorTableValuedFunction[] {
+    return source.tableValuedFunctions.map((fn, i) => {
+      addedTableValuedFunctions.set(fn, i);
+      return fn;
+    });
+  }
+
+  function translateParameters(source: TableProcessor): SerializedPartitionKey[] {
+    return source.parameters.map((key) => {
+      return { expr: visitExpr(replaceFunctionReferenceWithIndex, key.expr, null) };
+    });
   }
 
   function serializeDataSources(): SerializedDataSource[] {
@@ -49,7 +78,8 @@ export function serializeSyncPlan(plan: SyncPlan): SerializedSyncPlanUnstable {
         table: serializeTablePattern(source.sourceTable),
         outputTableName: source.outputTableName,
         filters: source.filters,
-        partitionBy: source.parameters,
+        tableValuedFunctions: serializeTableValued(source),
+        partitionBy: translateParameters(source),
         columns: source.columns
       } satisfies SerializedDataSource;
     });
@@ -63,8 +93,9 @@ export function serializeSyncPlan(plan: SyncPlan): SerializedSyncPlanUnstable {
         hash: source.hashCode,
         table: serializeTablePattern(source.sourceTable),
         filters: source.filters,
-        partitionBy: source.parameters,
-        output: source.outputs,
+        tableValuedFunctions: serializeTableValued(source),
+        partitionBy: translateParameters(source),
+        output: source.outputs.map((out) => visitExpr(replaceFunctionReferenceWithIndex, out, null)),
         lookupScope: source.defaultLookupScope
       } satisfies SerializedParameterIndexLookupCreator;
     });
@@ -147,37 +178,66 @@ export function deserializeSyncPlan(serialized: unknown): SyncPlan {
     throw new Error('Unknown sync plan version passed to deserializeSyncPlan()');
   }
 
-  function deserializeTablePattern(pattern: SerializedTablePattern): TablePattern {
-    return new TablePattern(`${pattern.connection}.${pattern.schema}`, pattern.table);
+  function deserializeTablePattern(pattern: SerializedTablePattern): ImplicitSchemaTablePattern {
+    if (pattern.schema) {
+      return new TablePattern(`${pattern.connection}.${pattern.schema}`, pattern.table);
+    } else {
+      return new ImplicitSchemaTablePattern(null, pattern.table);
+    }
+  }
+
+  let tableValuedFunctionsInScope: TableProcessorTableValuedFunction[] = [];
+
+  const replaceFunctionIndexWithReference = new MapSourceVisitor<
+    ColumnSqlParameterValue | SerializedTableProcessorTableValuedFunctionOutput,
+    ColumnSqlParameterValue | TableProcessorTableValuedFunctionOutput
+  >((value) => {
+    if ('function' in value) {
+      return { function: tableValuedFunctionsInScope[value.function], outputName: value.outputName };
+    } else {
+      return value;
+    }
+  });
+
+  function deserializeParameters(source: SerializedPartitionKey[]): PartitionKey[] {
+    return source.map((serializedKey) => {
+      return { expr: visitExpr(replaceFunctionIndexWithReference, serializedKey.expr, null) };
+    });
   }
 
   const plan = serialized as SerializedSyncPlanUnstable;
-  const dataSources = plan.dataSources.map((source) => {
+  const dataSources = plan.dataSources.map((source): StreamDataSource => {
+    const functions = (tableValuedFunctionsInScope = source.tableValuedFunctions);
+
     return {
       hashCode: source.hash,
       sourceTable: deserializeTablePattern(source.table),
+      tableValuedFunctions: functions,
       outputTableName: source.outputTableName,
       filters: source.filters,
-      parameters: source.partitionBy,
+      parameters: deserializeParameters(source.partitionBy),
       columns: source.columns
-    } satisfies StreamDataSource;
+    };
   });
-  const buckets = plan.buckets.map((bkt) => {
+  const buckets = plan.buckets.map((bkt): StreamBucketDataSource => {
     return {
       hashCode: bkt.hash,
       uniqueName: bkt.uniqueName,
       sources: bkt.sources.map((idx) => dataSources[idx])
-    } satisfies StreamBucketDataSource;
+    };
   });
-  const parameterIndexes = plan.parameterIndexes.map((source) => {
+  const parameterIndexes = plan.parameterIndexes.map((source): StreamParameterIndexLookupCreator => {
+    const functions = (tableValuedFunctionsInScope = source.tableValuedFunctions);
+
     return {
       hashCode: source.hash,
       sourceTable: deserializeTablePattern(source.table),
+      tableValuedFunctions: functions,
       filters: source.filters,
-      parameters: source.partitionBy,
-      outputs: source.output,
+      parameters: deserializeParameters(source.partitionBy),
+      outputs: source.output.map((out) => visitExpr(replaceFunctionIndexWithReference, out, null)),
       defaultLookupScope: source.lookupScope
-    } satisfies StreamParameterIndexLookupCreator;
+    };
   });
 
   function deserializeParameterValue(stages: ExpandingLookup[][], value: SerializedParameterValue): ParameterValue {
@@ -210,7 +270,7 @@ export function deserializeSyncPlan(serialized: unknown): SyncPlan {
           functionInputs: source.functionInputs,
           outputs: source.outputs,
           filters: source.filters
-        } satisfies EvaluateTableValuedFunction;
+        } satisfies EvaluateTableValuedFunction<RequestSqlParameterValue>;
     }
   }
 
@@ -263,9 +323,18 @@ interface SerializedBucketDataSource {
 }
 
 interface SerializedTablePattern {
-  connection: string;
-  schema: string;
+  connection: string | null;
+  schema: string | null;
   table: string;
+}
+
+interface SerializedTableProcessorTableValuedFunctionOutput {
+  function: number;
+  outputName: string;
+}
+
+interface SerializedPartitionKey {
+  expr: SqlExpression<ColumnSqlParameterValue | SerializedTableProcessorTableValuedFunctionOutput>;
 }
 
 interface SerializedDataSource {
@@ -274,16 +343,18 @@ interface SerializedDataSource {
   hash: number;
   columns: ColumnSource[];
   filters: SqlExpression<ColumnSqlParameterValue>[];
-  partitionBy: PartitionKey[];
+  tableValuedFunctions: TableProcessorTableValuedFunction[];
+  partitionBy: SerializedPartitionKey[];
 }
 
 interface SerializedParameterIndexLookupCreator {
   table: SerializedTablePattern;
   hash: number;
   lookupScope: ParameterLookupScope;
-  output: SqlExpression<ColumnSqlParameterValue>[];
+  output: SqlExpression<ColumnSqlParameterValue | SerializedTableProcessorTableValuedFunctionOutput>[];
   filters: SqlExpression<ColumnSqlParameterValue>[];
-  partitionBy: PartitionKey[];
+  tableValuedFunctions: TableProcessorTableValuedFunction[];
+  partitionBy: SerializedPartitionKey[];
 }
 
 interface SerializedStream {