+++ packages/sync-rules/test/src/sync_plan/evaluator/output_schema.test.ts
@@ -1,17 +1,17 @@
 import { describe, expect, test } from 'vitest';
 import {
-  addPrecompiledSyncPlanToRules,
   ColumnDefinition,
   sqlTypeName,
   CompatibilityContext,
   DEFAULT_TAG,
   javaScriptExpressionEngine,
   SourceTableDefinition,
-  SqlSyncRules,
   StaticSchema,
-  TablePattern
+  TablePattern,
+  PrecompiledSyncConfig,
+  deserializeSyncPlan
 } from '../../../../src/index.js';
-import { compileToSyncPlanWithoutErrors } from '../../compiler/utils.js';
+import { compileSingleStreamAndSerialize } from '../../compiler/utils.js';
 
 describe('schema inference', () => {
   const assetsTable: SourceTableDefinition = {
@@ -36,12 +36,13 @@ describe('schema inference', () => {
   ]);
 
   function generateSchema(...queries: string[]) {
-    const plan = compileToSyncPlanWithoutErrors([{ name: 'stream', queries }]);
-    const rules = new SqlSyncRules('');
-
-    addPrecompiledSyncPlanToRules(plan, rules, {
+    const serializedPlan = compileSingleStreamAndSerialize(...queries);
+    const plan = deserializeSyncPlan(serializedPlan);
+    const rules = new PrecompiledSyncConfig(plan, {
       // Engine isn't actually used here, but required to load sync plan
-      engine: javaScriptExpressionEngine(CompatibilityContext.FULL_BACKWARDS_COMPATIBILITY)
+      engine: javaScriptExpressionEngine(CompatibilityContext.FULL_BACKWARDS_COMPATIBILITY),
+      sourceText: '',
+      defaultSchema: 'test_schema'
     });
 
     const outputSchema: Record<string, Record<string, ColumnDefinition>> = {};