+++ packages/sync-rules/test/src/sync_plan/evaluator/utils.ts
@@ -1,20 +1,20 @@
 import {
   HydratedSyncRules,
-  SqlSyncRules,
   versionedHydrationState,
-  addPrecompiledSyncPlanToRules,
   javaScriptExpressionEngine,
   CompatibilityContext,
-  CompatibilityEdition
+  CompatibilityEdition,
+  SyncConfig,
+  PrecompiledSyncConfig
 } from '../../../../src/index.js';
-import { compileToSyncPlanWithoutErrors, SyncStreamInput } from '../../compiler/utils.js';
+import { compileToSyncPlanWithoutErrors } from '../../compiler/utils.js';
 import { test } from 'vitest';
 import { ScalarExpressionEngine } from '../../../../src/sync_plan/engine/scalar_expression_engine.js';
 
 interface SyncTest {
   engine: ScalarExpressionEngine;
-  prepareWithoutHydration(inputs: SyncStreamInput[]): SqlSyncRules;
-  prepareSyncStreams(inputs: SyncStreamInput[]): HydratedSyncRules;
+  prepareWithoutHydration(yaml: string): SyncConfig;
+  prepareSyncStreams(yaml: string): HydratedSyncRules;
 }
 
 export const syncTest = test.extend<{ sync: SyncTest }>({
@@ -25,10 +25,7 @@ export const syncTest = test.extend<{ sync: SyncTest }>({
       engine,
       prepareWithoutHydration: (inputs) => {
         const plan = compileToSyncPlanWithoutErrors(inputs);
-        const rules = new SqlSyncRules('');
-
-        addPrecompiledSyncPlanToRules(plan, rules, { engine });
-        return rules;
+        return new PrecompiledSyncConfig(plan, { engine, sourceText: '', defaultSchema: 'test_schema' });
       },
       prepareSyncStreams(inputs) {
         return this.prepareWithoutHydration(inputs).hydrate({ hydrationState: versionedHydrationState(1) });