+++ packages/service-core-tests/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @powersync/service-core-tests
 
+## 0.13.1
+
+### Patch Changes
+
+- 479997b: Introduce `BaseSyncConfig` to represent SQL-based sync rules and precompiled sync plans.
+- d1c2228: [MongoDB Storage] Optimize the compact job, avoiding re-compacting buckets in the same job.
+- Updated dependencies [0e99ce0]
+- Updated dependencies [479997b]
+- Updated dependencies [d1c2228]
+- Updated dependencies [1a1a4cc]
+  - @powersync/service-sync-rules@0.31.0
+  - @powersync/service-core@1.19.1
+
+## 0.13.0
+
+### Minor Changes
+
+- e578245: [Internal] Refactor sync rule representation to split out the parsed definitions from the hydrated state.
+
+### Patch Changes
+
+- 05b9661: Remove `parameterQueryLookups` from `BucketParameterQuerier`.
+- Updated dependencies [05b9661]
+- Updated dependencies [eaa04cc]
+- Updated dependencies [781d0e3]
+- Updated dependencies [e578245]
+- Updated dependencies [3040079]
+- Updated dependencies [3b2c512]
+- Updated dependencies [a02cc58]
+  - @powersync/service-core@1.19.0
+  - @powersync/service-sync-rules@0.30.0
+
 ## 0.12.16
 
 ### Patch Changes